supabase_discord_pdf_uploading/pyproject.toml
[project]
name = "supabase_py_discord_pdf_uploading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "supabase>=2.13.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]
